Took API call logic out of WeatherData and organized in CollectWeatherData class
