+++ Breathe/app/src/main/java/com/ybeltagy/breathe/data/WeatherData.java
@@ -49,102 +49,6 @@ public class WeatherData {
         setWeatherEPAIndex(aQI);
     }
 
-    //fixme: extract to secret file later.
-    //todo: use volley
-    //todo: refactor
-    //todo: use endtime.
-    //todo: better way to build the string.
-    //fixme: handle errors and exceptions.
-    //todo: comment
-    //todo: error handling for the UX
-    //todo: make calendar use UTC
-    //todo: Use a broadcast receiver.
-    public static String apiKey;
-
-    public static final String TEMPERATURE = "temperature";
-    public static final String HUMIDITY = "humidity";
-    public static final String EPAINDEX = "epaIndex";
-    public static final String PRECIPITATIONINTENSITY = "precipitationIntensity";
-    public static final String TREEINDEX = "treeIndex";
-    public static final String GRASSINDEX = "grassIndex";
-
-    /**
-     *
-     * @param calendar
-     * @return
-     */
-    public static String getTimestampISO8601(Calendar calendar){
-        Date date = calendar.getTime();
-        TimeZone tz = TimeZone.getTimeZone("UTC");
-        DateFormat df = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'"); // Quoted "Z" to indicate UTC, no timezone offset
-        df.setTimeZone(tz);
-        return df.format(date);
-    }
-
-    /**
-     *
-     * @return
-     */
-    public static WeatherData syncGetWeatherData(Calendar startTime, double latitude, double longitude){
-
-        Calendar endTime = (Calendar) startTime.clone();
-        endTime.add(Calendar.SECOND, 60);
-
-        String url = "https://data.climacell.co/v4/timelines?location=" + latitude + "," + longitude +
-                "&fields=" + TEMPERATURE + "," + HUMIDITY + "," + EPAINDEX + "," + PRECIPITATIONINTENSITY + "," + TREEINDEX + "," + GRASSINDEX +
-                "&startTime=" + getTimestampISO8601(startTime) +
-                "&endTime=" + getTimestampISO8601(endTime) +
-                "&timesteps=1m" +
-                "&apikey=" + apiKey;
-
-        Log.d("WeatherData", url);
-        OkHttpClient client = new OkHttpClient();
-
-        Request request = new Request.Builder()
-                .url(url)
-                .build();
-
-        Log.d("WeatherData", request.toString());
-        // FIXME: (after merge) change JSON_Object.getDouble(), .getInt(), etc. to .OptDouble(),
-        //       .OptInt() and check for null.  If those values are null for the object, assign the
-        //       WeatherData fields to the null constant values
-        try (Response response = client.newCall(request).execute()) {
-            JSONObject obj = new JSONObject(response.body().string()).getJSONObject("data");
-            obj = (obj.getJSONArray("timelines").getJSONObject(0)).getJSONArray("intervals").getJSONObject(0).getJSONObject("values");
-
-            WeatherData weatherData = new WeatherData();
-
-            weatherData.weatherTemperature = (float) obj.getDouble(TEMPERATURE);
-            Log.d("WeatherData", "Temperature : " + weatherData.weatherTemperature);
-
-            weatherData.weatherHumidity = (float) obj.getDouble(HUMIDITY);
-            Log.d("WeatherData", "Humidity : " + weatherData.weatherHumidity);
-
-            weatherData.weatherEPAIndex = obj.getInt(EPAINDEX);
-            Log.d("WeatherData", "EPA Index : " + weatherData.weatherEPAIndex);
-
-            weatherData.weatherPrecipitationIntensity
-                    = (float) obj.getDouble(PRECIPITATIONINTENSITY);
-            Log.d("WeatherData",
-                    "Precipitation Intensity : " + weatherData.weatherPrecipitationIntensity);
-
-            weatherData.weatherGrassIndex = Level.values()[obj.getInt(GRASSINDEX)];
-            Log.d("WeatherData", "Grass Index : " + weatherData.weatherGrassIndex);
-
-            weatherData.weatherTreeIndex = Level.values()[obj.getInt(TREEINDEX)];
-            Log.d("WeatherData", "Tree Index : " + weatherData.weatherTreeIndex);
-
-            return weatherData;
-
-        }catch(IOException e){
-            Log.e("WeatherData", "IOException : " + e.getMessage());
-            return null;
-        }catch(JSONException e){
-            Log.e("WeatherData", "JSONException : " + e.getMessage());
-            return null;
-        }
-
-    }
 
     //todo: implement a toString method.
